+++ adbai-20250812/alibabacloud_adbai20250812/models/_create_embodied_aiplatform_shrink_request.py
@@ -8,6 +8,7 @@ class CreateEmbodiedAIPlatformShrinkRequest(DaraModel):
     def __init__(
         self,
         dbcluster_id: str = None,
+        device_count: int = None,
         platform_name: str = None,
         ray_config_shrink: str = None,
         region_id: str = None,
@@ -15,6 +16,7 @@ class CreateEmbodiedAIPlatformShrinkRequest(DaraModel):
     ):
         # This parameter is required.
         self.dbcluster_id = dbcluster_id
+        self.device_count = device_count
         # This parameter is required.
         self.platform_name = platform_name
         self.ray_config_shrink = ray_config_shrink
@@ -32,6 +34,9 @@ class CreateEmbodiedAIPlatformShrinkRequest(DaraModel):
             result = _map
         if self.dbcluster_id is not None:
             result['DBClusterId'] = self.dbcluster_id
+
+        if self.device_count is not None:
+            result['DeviceCount'] = self.device_count
 
         if self.platform_name is not None:
             result['PlatformName'] = self.platform_name
@@ -51,6 +56,9 @@ class CreateEmbodiedAIPlatformShrinkRequest(DaraModel):
         m = m or dict()
         if m.get('DBClusterId') is not None:
             self.dbcluster_id = m.get('DBClusterId')
+
+        if m.get('DeviceCount') is not None:
+            self.device_count = m.get('DeviceCount')
 
         if m.get('PlatformName') is not None:
             self.platform_name = m.get('PlatformName')